+++ tailwind.config.js
@@ -37,6 +37,7 @@ export default {
 				"184": "184px",
 				"220": "220px",
 				"282": "282px",
+				"1200": "1200px",
 			},
 			margin: {
 				"74": "74px",
